requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.10.4
pydantic-settings==2.4.0
email-validator>=2.0.0
gunicorn

# LLM Integration
openai>=1.58.1,<2
anthropic==0.18.1
langchain==0.3.6
langchain-community==0.3.4
langchain-openai==0.2.14
langchain-neo4j==0.1.0

# Machine Learning
scikit-learn==1.4.2

# Agent Framework
langgraph==0.2.60
langgraph-checkpoint==2.0.8

# LightRAG - Lightweight GraphRAG Alternative
lightrag-hku>=0.0.3

# Vector Database & Embeddings
pymilvus==2.3.7

# Database
sqlalchemy==2.0.25
pymysql==1.1.0
redis==5.0.1
neo4j==5.27.0

# HTTP Client
httpx==0.26.0
requests==2.31.0
bcrypt==4.1.2
# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
pyyaml==6.0.1
jinja2==3.1.3
pyahocorasick==2.0.0

# Web Scraping & Data Extraction
beautifulsoup4==4.12.3
lxml==5.1.0
fake-useragent==1.4.0
playwright==1.40.0

# Logging & Monitoring
loguru==0.7.2
prometheus-client==0.19.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0
pillow
IPython
matplotlib
seaborn